pyproject.toml
[tool.poetry]
name = "pythonproject3"
version = "0.1.0"
description = ""
authors = ["worker76 <worker76@bwg-io.site>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiopg = "^1.4.0"
pydantic = "^1.10.8"
dependency-injector = "^4.41.0"
yoyo-migrations = "^8.2.0"
python-dotenv = "^1.0.0"
aioredis = "^2.0.1"
starlette = "^0.27.0"
fastapi = "^0.96.0"
redis = "^4.5.5"
pillow = "^9.5.0"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
isort = "^5.10.1"
black = "^22.1.0"
docformatter = "^1.4"
add-trailing-comma = "^2.2.1"
flake8 = "^4.0.1"
mypy = "^0.931"
safety = "^1.10.3"
bandit = "^1.7.1"
itsdangerous = "^2.0.1"
MarkupSafe = "^2.1.0"
Sphinx = "^4.4.0"
sphinx-pydantic = "^0.1.1"
sphinx-rtd-theme = "^1.0.0"
pytest-asyncio = "^0.18.2"
pytest = "^7.0.1"
coverage = "^6.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
